move proto functions into protobuilder
